test(plugins): drop the dynamic client suite for ORM-free discovery tests
The convergence it claimed to cover lives in
plugin_oauth_clients_issuer_redirect_unique, which the mocked chain().limit()
could never observe; what it exercised was the in-process registering map.
Mocking drizzle-orm, the db client and the schema process-wide made it a flake
source for every file bun ordered after it.

discovery.test.ts keeps the assertions that need no database: the
first-party trust gate, a missing mcp url, a non-https authorization server,
an http authorization_endpoint, and metadata describing a foreign issuer. It
stubs global.fetch and restores it.
